Fixed compiler warnings in Pitch- and RollModel

+++ source/ApplicationLayer/Models/RollModel.cpp
@@ -33,5 +33,5 @@ void ApplicationLayer::Models::RollModel::Update(uint32_t)
 	float yAccel = m_YAxis.GetValue() / 1000.0f;
 	float zAccel = m_ZAxis.GetValue() / 1000.0f;
 
-    m_RollMilliDegrees = ((std::atan2(-yAccel, zAccel) * 180.0f) / Common::Math::PI) * 1000.0f;
+    m_RollMilliDegrees = static_cast<int32_t>(((std::atan2(-yAccel, zAccel) * 180.0f) / Common::Math::PI) * 1000.0f);
 }
